Move progress bar to separate file
